fix: Address CodeRabbit review feedback and improve code quality
- Fix Scene Manager memory leak by updating _gameElements array in remove()
- Extract shared setupNetworkRoutes function to tests/utils/network-setup.js
- Add error handling consistency for GameOverScreen initialization
- Add camera validation in SceneManager.setCamera()
- Remove debug console logs with [UPDATED VERSION] markers
- Remove unused _showNextText() method from SequenceManager
- Remove redundant AudioContext handling in startGame()
- Fix markdown formatting issues in documentation
- Clean up unnecessary blank lines in HTML
