+++ docs/20_design/sequence-design.md
@@ -9,10 +9,10 @@
 *   **タイトル表示:** 暗転した画面に、テキストで「忘れ去られた王国」というタイトルがフェードインで表示される。
 *   **風景の提示:** ゆっくりとフェードインしながら、荒廃した王国の風景（Three.jsのシーン）が映し出される。カメラはテキストに合わせてゆっくりとパン・ズームする。
 *   **ナレーション（テキスト）:**
-    *   「かつて、この王国は光に満ちていた。」
-    *   「しかし、謎の呪いにより、その輝きは失われた。」
-    *   「今、一匹の猫が目覚める。失われた記憶と、王国を救う使命を胸に。」
-    *   「希望の光は、再びこの地に灯るのか。」
+  *   「かつて、この王国は光に満ちていた。」
+  *   「しかし、謎の呪いにより、その輝きは失われた。」
+  *   「今、一匹の猫が目覚める。失われた記憶と、王国を救う使命を胸に。」
+  *   「希望の光は、再びこの地に灯るのか。」
 *   **ロゴ表示:** 「もふもふアドベンチャー」のロゴが画面中央に表示される。
 *   **BGM:** 荘厳で神秘的なBGMが流れる。
 
@@ -31,8 +31,8 @@
 *   **王国の再生:** 荒廃していた王国に光が差し込み、枯れていた木々に緑が戻り、花が咲き乱れる様子が映し出される。カメラは広範囲をゆっくりと俯瞰する。
 *   **主人公の姿:** 王国の中心で、主人公が静かにたたずみ、遠くを見つめる姿。
 *   **メッセージ（テキスト）:**
-    *   「呪いは解かれ、王国に平和が戻った。」
-    *   「しかし、冒険はまだ始まったばかりだ…」
+  *   「呪いは解かれ、王国に平和が戻った。」
+  *   「しかし、冒険はまだ始まったばかりだ…」
 *   **スタッフロール:** 簡易的なスタッフロール（開発者名、使用ライブラリなど）が下から上へゆっくりとスクロールする。
 *   **最終表示:** 「Fin」または「The End」が画面中央に表示される。
 *   **BGM:** 感動的で希望に満ちたBGMが流れる。
