+++ docs/40_technical/architecture.md
@@ -102,11 +102,11 @@ src/
 
 ### 8.1. カットシーン（シーケンス演出）の実現方法
 
-*   **採用方針:** Three.jsによるリアルタイムレンダリングで実装。AIによるコード生成とゲームの没入感を最大化するため。
-*   **実装詳細:**
-    *   `Game` クラス内にシーケンス演出を管理するメソッド（`playOpeningSequence()`, `playEndingSequence()`など）を実装。
-    *   各シーケンスは、Three.jsのシーン内でカメラワーク、オブジェクトの配置・アニメーション、テキスト表示などを組み合わせることで実現する。
-    *   シーケンス中は、`Game.gameState` を新しい状態（例: `GameState.SEQUENCE`）に設定し、プレイヤーの入力や通常のゲームロジックの更新を停止する。
-    *   シーケンスの進行は、時間経過や特定のアニメーション完了イベントに基づいて制御する。
-    *   シーケンス終了後、ゲーム状態を元に戻し、次の画面へ遷移させる。
+- **採用方針:** Three.jsによるリアルタイムレンダリングで実装。AIによるコード生成とゲームの没入感を最大化するため。
+- **実装詳細:**
+  - `Game` クラス内にシーケンス演出を管理するメソッド（`playOpeningSequence()`, `playEndingSequence()`など）を実装。
+  - 各シーケンスは、Three.jsのシーン内でカメラワーク、オブジェクトの配置・アニメーション、テキスト表示などを組み合わせることで実現する。
+  - シーケンス中は、`Game.gameState` を新しい状態（例: `GameState.SEQUENCE`）に設定し、プレイヤーの入力や通常のゲームロジックの更新を停止する。
+  - シーケンスの進行は、時間経過や特定のアニメーション完了イベントに基づいて制御する。
+  - シーケンス終了後、ゲーム状態を元に戻し、次の画面へ遷移させる。
