+++ public/index.html
@@ -11,8 +11,6 @@
     <link rel="stylesheet" href="styles/style.css" />
   </head>
   <body>
-    
-    
     <script type="importmap">
       {
         "imports": {
