+++ public/src/main.js
@@ -4,7 +4,7 @@ console.log('main.js module loaded successfully!');
 
 // Initialize and start the game
 try {
-  console.log('Creating Game instance... [UPDATED VERSION]');
+  console.log('Creating Game instance...');
   const game = new Game();
   console.log('Game instance created successfully');
   window.game = game; // Make game accessible globally for testing
