+++ tests/debug.spec.js
@@ -1,31 +1,5 @@
 const { test, expect } = require('@playwright/test');
-
-async function setupNetworkRoutes(page) {
-  await page.route(
-    'https://unpkg.com/three@0.160.0/build/three.module.js',
-    (route) => {
-      route.fulfill({
-        path: require('path').join(
-          __dirname,
-          '../node_modules/three/build/three.module.js'
-        ),
-      });
-    }
-  );
-  await page.route(
-    'https://unpkg.com/three@0.160.0/examples/jsm/**', 
-    (route) => {
-      const url = route.request().url();
-      const jsmPath = url.substring(url.indexOf('/jsm/') + 5);
-      const localPath = require('path').join(
-        __dirname,
-        '../node_modules/three/examples/jsm/',
-        jsmPath
-      );
-      route.fulfill({ path: localPath });
-    }
-  );
-}
+const { setupNetworkRoutes } = require('./utils/network-setup');
 
 test.describe('Debug Tests', () => {
   test('debug page load and game initialization', async ({ page }) => {
